+++ docs/release_notes.rst
@@ -10,10 +10,10 @@ What's new in 0.9.0dev0
 Features:
 ^^^^^^^^^
 
-* Introducing Schedulers to enable adaptive mutation and crossover probabilities;
+* Introducing Adaptive Schedulers to enable adaptive mutation and crossover probabilities;
   currently, supported schedulers are:
-  - :class:`~sklearn_genetic.schedules.ExponentialDecay`
-  - :class:`~sklearn_genetic.schedules.InverseDecay`
+  - :class:`~sklearn_genetic.schedules.ExponentialAdapter`
+  - :class:`~sklearn_genetic.schedules.InverseAdapter`
 
 * Changed the default values of `mutation_probability` and `crossover_probability`
   to 0.8 and 0.2, respectively.